[DCGEO-218] Refactor to use BigInteger as backing counter

+++ apps/bfd-common/bfd-common-util/src/test/java/gov/cms/bfd/common/generators/token/TokenPatternParserTest.java
@@ -66,7 +66,7 @@ public class TokenPatternParserTest {
 
     TokenPattern tokenPattern = parser.parse(pattern);
 
-    long tokenValue = tokenPattern.parseTokenValue("B2");
+    long tokenValue = tokenPattern.parseTokenValue("B2").longValue();
 
     assertEquals(6, tokenValue);
   }
